requirements.txt
# Core
python-telegram-bot==21.9
aiosqlite==0.20.0
python-dotenv==1.0.1
openai>=1.0.0

# Scheduling
APScheduler==3.10.4
pytz==2024.1

# HTTP
aiohttp==3.11.11
httpx==0.27.0
aiofiles==24.1.0

# Telethon (promo agent only)
telethon==1.38.1

# Utilities
python-dateutil==2.9.0
